requirements.txt
aiohttp>=3.11.10
authlib>=1.3.2
av>=13.1.0
google-adk>=1.19.0
google-cloud-logging>=3.11.3
google-cloud-secret-manager>=2.21.1
google-cloud-speech>=2.28.0
google-cloud-texttospeech>=2.18.0
google-genai>=1.0.0
jinja2>=3.1.4
matplotlib>=3.9.2
numpy>=2.1.3
pandas>=2.2.3
pillow>=11.0.0
plotly>=5.24.1
prometheus-client>=0.21.0
psycopg2-binary>=2.9.10
pydantic>=2.10.3
pytest>=8.3.4
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
python-jose>=3.3.0
reportlab>=4.2.5
requests>=2.32.3
restrictedpython>=7.4
scikit-learn>=1.6.0
scipy>=1.14.1
sentry-sdk>=2.18.0
sqlalchemy>=2.0.36
streamlit>=1.51.0
trafilatura>=1.12.2
yfinance>=0.2.48
